+++ R/disp_DA.R
@@ -209,7 +209,7 @@ disp_DA <- function(subfreq,
 #' @author Lukas Soenning
 #' 
 #' @details 
-#' This function takes as input a term-document matrix and returns, for each item (i.e. each row) the dispersion measure \eqn{D_{A}}. The rows in the matrix represent the items, and the columns the corpus parts. Importantly, the term-document matrix must include an additional row that records the size of the corpus parts. For a proper term-document matrix, which includes all items that appear in the corpus, this can be added as a column margin, which sums the frequencies in each column. If the matrix only includes a selection of items drawn from the corpus, this information cannot be derived from the matrix and must be provided as a separate row.
+#' This function takes as input a term-document matrix and returns, for each item (i.e. each row) the dispersion measure \eqn{D_{A}}. The rows in the input matrix represent the items, and the columns the corpus parts. Importantly, the term-document matrix must include an additional row that records the size of the corpus parts. For a proper term-document matrix, which includes all items that appear in the corpus, this can be added as a column margin, which sums the frequencies in each column. If the matrix only includes a selection of items drawn from the corpus, this information cannot be derived from the matrix and must be provided as a separate row.
 #' 
 #' - Directionality: \eqn{D_{A}} ranges from 0 to 1. The conventional scaling of dispersion measures (see Juilland & Chang-Rodriguez 1964; Carroll 1970; Rosengren 1971) assigns higher values to more even/dispersed/balanced distributions of subfrequencies across corpus parts. This is the default. Gries (2008) uses the reverse scaling, with higher values denoting a more uneven/bursty/concentrated distribution; use `directionality = 'gries'` to choose this option.
 #' 
@@ -242,8 +242,9 @@ disp_DA <- function(subfreq,
 #' The value `shortcut_mod` adds a minor modification to the computational shortcut to ensure \eqn{D_{A}} does not exceed 1 (on the conventional dispersion scale):
 #' 
 #'    \eqn{\frac{2\left(\sum_{i = 1}^{k} (i \times r_i^{sorted}) - 1\right)}{k-1} \times \frac{k}{k - 1}}
+#'
 #' 
-#' @returns A numeric vector the same length as the number of items in the term-document matrix
+#' @returns A data frame with one row per item
 #' 
 ##' @references 
 #' 
@@ -283,6 +284,7 @@ disp_DA_tdm <- function(tdm,
                         procedure = "basic",
                         freq_adjust = FALSE,
                         freq_adjust_method = "even",
+                        add_frequency = TRUE,
                         unit_interval = TRUE,
                         digits = NULL,
                         verbose = TRUE,
@@ -453,6 +455,24 @@ disp_DA_tdm <- function(tdm,
   output <- t(output)
   
   if (!is.null(digits)) output <- round(output, digits)
+  
+  output <- data.frame(
+    item = colnames(output),
+    DP = as.numeric(output)
+  )
+  
+  if (freq_adjust == TRUE){
+    colnames(output)[2] <- "DP_nofreq"
+  }
+  
+  if (add_frequency == TRUE){
+    if (row_partsize == "first"){
+      output$frequency <- rowSums(tdm[-1,])
+    }
+    if (row_partsize == "last"){
+      output$frequency <- rowSums(tdm[-nrow(tdm),])
+    }
+  }
   
   if (print_scores != FALSE) print(output)